+++ orchestrator/cosmos/network.go
@@ -2,59 +2,22 @@ package cosmos
 
 import (
 	"context"
-	"fmt"
-	"github.com/InjectiveLabs/peggo/orchestrator/cosmos/tendermint"
-	peggytypes "github.com/InjectiveLabs/sdk-go/chain/peggy/types"
-	"github.com/InjectiveLabs/sdk-go/client/chain"
-	comethttp "github.com/cometbft/cometbft/rpc/client/http"
-	"github.com/pkg/errors"
-	log "github.com/xlab/suplog"
 	"time"
 
+	peggytypes "github.com/InjectiveLabs/sdk-go/chain/peggy/types"
+	"github.com/InjectiveLabs/sdk-go/client/chain"
+	clientcommon "github.com/InjectiveLabs/sdk-go/client/common"
+	comethttp "github.com/cometbft/cometbft/rpc/client/http"
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
+	"github.com/pkg/errors"
+	log "github.com/xlab/suplog"
+	"google.golang.org/grpc"
+	"google.golang.org/grpc/connectivity"
 
 	"github.com/InjectiveLabs/peggo/orchestrator/cosmos/peggy"
+	"github.com/InjectiveLabs/peggo/orchestrator/cosmos/tendermint"
 	"github.com/InjectiveLabs/peggo/orchestrator/ethereum/keystore"
-	clientcommon "github.com/InjectiveLabs/sdk-go/client/common"
 )
-
-type NetworkConfig struct {
-	ChainID,
-	ValidatorAddress,
-	CosmosGRPC,
-	TendermintRPC,
-	GasPrice string
-}
-
-func (cfg NetworkConfig) loadClientConfig() clientcommon.Network {
-	if custom := cfg.CosmosGRPC != "" && cfg.TendermintRPC != ""; custom {
-		c := clientcommon.LoadNetwork("devnet", "")
-		c.Name = "custom"
-		c.ChainId = cfg.ChainID
-		c.Fee_denom = "inj"
-		c.TmEndpoint = cfg.TendermintRPC
-		c.ChainGrpcEndpoint = cfg.CosmosGRPC
-		c.ExplorerGrpcEndpoint = ""
-		c.LcdEndpoint = ""
-		c.ExplorerGrpcEndpoint = ""
-
-		return c
-	}
-
-	var networkName string
-	switch cfg.ChainID {
-	case "injective-1":
-		networkName = "mainnet"
-	case "injective-777":
-		networkName = "devnet"
-	case "injective-888":
-		networkName = "testnet"
-	default:
-		panic(fmt.Errorf("no provider for chain id %s", cfg.ChainID))
-	}
-
-	return clientcommon.LoadNetwork(networkName, "lb")
-}
 
 type Network interface {
 	GetBlockTime(ctx context.Context, height int64) (time.Time, error)
@@ -82,8 +45,7 @@ func NewCosmosNetwork(k keyring.Keyring, ethSignFn keystore.PersonalSignFn, cfg 
 		return nil, errors.Wrapf(err, "failed to connect to Tendermint RPC %s", clientCfg.TmEndpoint)
 	}
 
-	clientCtx = clientCtx.WithNodeURI(clientCfg.TmEndpoint)
-	clientCtx = clientCtx.WithClient(tmRPC)
+	clientCtx = clientCtx.WithNodeURI(clientCfg.TmEndpoint).WithClient(tmRPC)
 
 	chainClient, err := chain.NewChainClient(clientCtx, clientCfg, clientcommon.OptionGasPrices(cfg.GasPrice))
 	if err != nil {
@@ -92,15 +54,10 @@ func NewCosmosNetwork(k keyring.Keyring, ethSignFn keystore.PersonalSignFn, cfg 
 
 	time.Sleep(1 * time.Second)
 
-	daemonWaitCtx, cancelWait := context.WithTimeout(context.Background(), time.Minute)
-	defer cancelWait()
-
-	grpcConn := chainClient.QueryClient()
-	waitForService(daemonWaitCtx, grpcConn)
-	peggyQuerier := peggytypes.NewQueryClient(grpcConn)
+	conn := awaitConnection(1*time.Minute, chainClient)
 
 	n := network{
-		QueryClient:     peggy.NewQueryClient(peggyQuerier),
+		QueryClient:     peggy.NewQueryClient(peggytypes.NewQueryClient(conn)),
 		BroadcastClient: peggy.NewBroadcastClient(chainClient, ethSignFn),
 		Client:          tendermint.NewRPCClient(clientCfg.TmEndpoint),
 	}
@@ -108,9 +65,9 @@ func NewCosmosNetwork(k keyring.Keyring, ethSignFn keystore.PersonalSignFn, cfg 
 	log.WithFields(log.Fields{
 		"chain_id":   cfg.ChainID,
 		"addr":       cfg.ValidatorAddress,
-		"injective":  cfg.CosmosGRPC,
+		"injective":  clientCfg.ChainGrpcEndpoint,
 		"tendermint": clientCfg.TmEndpoint,
-	}).Infoln("connected to custom Injective endpoints")
+	}).Infoln("connected to Injective network")
 
 	return n, nil
 }
@@ -122,4 +79,47 @@ func (n network) GetBlockTime(ctx context.Context, height int64) (time.Time, err
 	}
 
 	return block.Block.Time, nil
+}
+
+// waitForService awaits an active ClientConn to a GRPC service.
+func waitForService(ctx context.Context, clientConn *grpc.ClientConn) {
+	for {
+		select {
+		case <-ctx.Done():
+			log.Fatalln("GRPC service wait timed out")
+		default:
+			state := clientConn.GetState()
+
+			if state != connectivity.Ready {
+				log.WithField("state", state.String()).Warningln("state of GRPC connection not ready")
+				time.Sleep(5 * time.Second)
+				continue
+			}
+
+			return
+		}
+	}
+}
+
+func awaitConnection(timeout time.Duration, client chain.ChainClient) *grpc.ClientConn {
+	ctx, cancelWait := context.WithTimeout(context.Background(), timeout)
+	defer cancelWait()
+
+	grpcConn := client.QueryClient()
+
+	for {
+		select {
+		case <-ctx.Done():
+			log.Fatalln("GRPC service wait timed out")
+		default:
+			state := grpcConn.GetState()
+			if state != connectivity.Ready {
+				log.WithField("state", state.String()).Warningln("state of GRPC connection not ready")
+				time.Sleep(5 * time.Second)
+				continue
+			}
+
+			return grpcConn
+		}
+	}
 }
